stremlit_sample_voicememo/requirements.txt
streamlit
pyaudio
websockets
websocket-client
asyncio
numpy
wave
uuid
whisper
soundfile
scipy 